requirements.txt
argon2==0.1.10
bcrypt==3.2.2
Bcrypt-Flask==1.0.2
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.0
dropbox==11.36.2
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==20.0.4
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
ply==3.11
psycopg2==2.9.9
pycparser==2.21
python-dotenv==1.0.1
python-engineio==4.9.0
pytz==2024.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.27
stone==3.3.3
typing_extensions==4.9.0
urllib3==2.2.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
